% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{get_translations}
\alias{get_translations}
\title{Obtenir les traductions de l'application}
\usage{
get_translations()
}
\value{
Une liste de traductions
}
\description{
Obtenir les traductions de l'application
}
