% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R, R/visualizations.R
\name{load_whatsapp_data}
\alias{load_whatsapp_data}
\title{Charger les données WhatsApp}
\usage{
load_whatsapp_data(file_path)

load_whatsapp_data(file_path)
}
\arguments{
\item{file_path}{Chemin vers le fichier WhatsApp}
}
\value{
Un dataframe contenant les données WhatsApp

Un dataframe avec les données WhatsApp
}
\description{
Charger les données WhatsApp

Charger les données WhatsApp
}
