% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R, R/visualizations.R
\name{plot_avg_message_length}
\alias{plot_avg_message_length}
\title{Créer un graphique de la longueur moyenne des messages}
\usage{
plot_avg_message_length(data, selected_authors = NULL)

plot_avg_message_length(data, selected_authors = NULL)
}
\arguments{
\item{data}{Données WhatsApp}

\item{selected_authors}{Liste des auteurs sélectionnés}
}
\value{
Un graphique plotly

Un graphique plotly
}
\description{
Créer un graphique de la longueur moyenne des messages

Créer un graphique de la longueur moyenne des messages
}
