% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R, R/visualizations.R
\name{plot_lexical_diversity}
\alias{plot_lexical_diversity}
\title{Créer un graphique de la diversité lexicale}
\usage{
plot_lexical_diversity(data, selected_authors = NULL)

plot_lexical_diversity(data, selected_authors = NULL)
}
\arguments{
\item{data}{Données WhatsApp}

\item{selected_authors}{Liste des auteurs sélectionnés}
}
\value{
Un graphique plotly

Un graphique plotly
}
\description{
Créer un graphique de la diversité lexicale

Créer un graphique de la diversité lexicale
}
