% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R, R/visualizations.R
\name{plot_message_evolution}
\alias{plot_message_evolution}
\title{Créer un graphique de l'évolution temporelle des messages}
\usage{
plot_message_evolution(data, selected_authors = NULL)

plot_message_evolution(data, selected_authors = NULL)
}
\arguments{
\item{data}{Données WhatsApp}

\item{selected_authors}{Liste des auteurs sélectionnés}
}
\value{
Un graphique plotly

Un graphique plotly
}
\description{
Créer un graphique de l'évolution temporelle des messages

Créer un graphique de l'évolution temporelle des messages
}
