% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R, R/visualizations.R
\name{plot_most_used_words}
\alias{plot_most_used_words}
\title{Créer un graphique des mots les plus utilisés}
\usage{
plot_most_used_words(data, selected_authors = NULL)

plot_most_used_words(data, selected_authors = NULL)
}
\arguments{
\item{data}{Données WhatsApp}

\item{selected_authors}{Liste des auteurs sélectionnés}
}
\value{
Un graphique plotly

Un graphique plotly
}
\description{
Créer un graphique des mots les plus utilisés

Créer un graphique des mots les plus utilisés
}
