% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{update_ui_translations}
\alias{update_ui_translations}
\title{Mettre à jour les traductions de l'interface}
\usage{
update_ui_translations(session, translations)
}
\arguments{
\item{session}{Session Shiny}

\item{translations}{Traductions à appliquer}
}
\description{
Mettre à jour les traductions de l'interface
}
